update meeting lat long
The latitude and longitude are for the initial general location selection. Search radius is also included.

+++ src/server/app/models/meeting.js
@@ -4,8 +4,8 @@ module.exports = function (sequelize, DataTypes) {
     title: { type: DataTypes.STRING, allowNull: false },
     closeoutTime: { type: DataTypes.DATE, allowNull: false},
     generalLocation: { type: DataTypes.STRING, defaultValue: "", allowNull: false }, //  general location
-    latitude: { type: DataTypes.INTEGER, defaultValue: null, validate: { min: -90, max: 90 }}, // Location latitude
-    longitude: { type: DataTypes.INTEGER, defaultValue: null, validate: { min: -180, max: 180 }}, // Location longitude
+    latitude: { type: DataTypes.DOUBLE, defaultValue: null, validate: { min: -90, max: 90 }}, // Location latitude - general meeting area
+    longitude: { type: DataTypes.DOUBLE, defaultValue: null, validate: { min: -180, max: 180 }}, // Location longitude - general meeting area
     radius: { type: DataTypes.INTEGER, defaultValue: 1000, allowNull: false }, // Location Search Radius in meters
     duration: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 60 }, // Meeting durating - units in minutes -- default 1 hour
     invited: { type: DataTypes.JSON, defaultValue: "", allowNull: false }, // List of invited individuals (emails)
